Newchanges

+++ WebApplication1/Controllers/OrderController.cs
@@ -15,26 +15,69 @@ namespace WebApplication1.Controllers
     {
         private readonly IOrderService _orderService;
         private readonly ServiceContext _serviceContext;
+        //private readonly IDetallePedidoService _detallePedidoService;
 
         public OrderController(IOrderService orderService, ServiceContext serviceContext)
         {
             _orderService = orderService;
             _serviceContext = serviceContext;
+            //_detallePedidoService = detallePedidoService;
         }
 
-        [HttpPost(Name = "insertOrder")]
-        public IActionResult Post([FromQuery] string userNombreUsuario, [FromQuery] string userContraseña, [FromBody] OrderItem orderItem)
+        [HttpPost (Name = "InsertOrder")]
+        //public ActionResult CreateOrderWithDetalle([FromBody] OrderItem orderItem, [FromHeader] string userNombreUsuario, [FromHeader] string userContraseña)
+        //{
+        //    var existingUser = _serviceContext.Set<UserItem>()
+        //.Where(u => u.NombreUsuario == userNombreUsuario && u.Contraseña == userContraseña && u.Rol == 1)
+        //.FirstOrDefault();
+
+        //    if (existingUser == null)
+        //    {
+        //        return Unauthorized();
+        //    }
+        //    using (var transaction = _serviceContext.Database.BeginTransaction())
+        //    {
+        //        try
+        //        {
+        //            var orderId = _orderService.insertOrder(orderItem);
+        //            //_detallePedidoService.InsertDetalleWithOrder(detallePedidoItem, orderId);
+
+        //            transaction.Commit();
+        //            return Ok("Pedido y detalle agregados exitosamente.");
+        //        }
+        //        catch (Exception ex)
+        //        {
+        //            transaction.Rollback();
+        //            return BadRequest($"Error: {ex.Message}");
+        //        }
+        //    }
+        //}
+
+        public int InsertOrder([FromBody] OrderItem orderItem, [FromQuery] string userNombreUsuario, [FromQuery] string userContraseña)
         {
-            // ... lógica de autenticación ...
+            var existingUser = _serviceContext.Set<UserItem>()
+        .Where(u => u.NombreUsuario == userNombreUsuario && u.Contraseña == userContraseña && u.Rol == 1)
+        .FirstOrDefault();
 
-            int orderId = _orderService.insertOrder(orderItem);
-            return Ok(orderId);
+            if (existingUser != null)
+            {
+                return _orderService.insertOrder(orderItem);
+            }
+            else
+            {
+                throw new InvalidCredentialException("El ususario no esta autorizado o no existe");
+            }
+        
+           
         }
+
+
+
+
 
 
         [HttpPut("{id}", Name = "UpdateOrders")]
-
-        public IActionResult Put(int id, [FromHeader] string userNombreUsuario, [FromHeader] string userContraseña, [FromBody] OrderItem updatedOrdertItem)
+        public IActionResult Put(int id, [FromHeader] string userNombreUsuario, [FromHeader] string userContraseña, [FromBody] OrderItem updatedOrderItem)
         {
             var existingUser = _serviceContext.Set<UserItem>()
                 .Where(u => u.NombreUsuario == userNombreUsuario && u.Contraseña == userContraseña && u.Rol == 1)
@@ -42,7 +85,7 @@ namespace WebApplication1.Controllers
 
             if (existingUser == null)
             {
-                return Unauthorized(); // Cambiamos NotFound a Unauthorized para ocultar que existe o no el usuario.
+                return Unauthorized();
             }
 
             var existingOrderItem = _serviceContext.Set<OrderItem>()
@@ -55,9 +98,9 @@ namespace WebApplication1.Controllers
             }
             else
             {
-                existingOrderItem.OrderDate = updatedOrdertItem.OrderDate;
-                existingOrderItem.Delivered = updatedOrdertItem.Delivered;
-                existingOrderItem.Charged = updatedOrdertItem.Charged;
+                existingOrderItem.OrderDate = updatedOrderItem.OrderDate;
+                existingOrderItem.Delivered = updatedOrderItem.Delivered;
+                existingOrderItem.Charged = updatedOrderItem.Charged;
             }
 
             _orderService.UpdateOrder(existingOrderItem);
@@ -65,12 +108,45 @@ namespace WebApplication1.Controllers
         }
 
         [HttpDelete("{id}", Name = "DeleteOrder")]
-        public IActionResult Delete(int id)
+        public IActionResult Delete(int id, [FromHeader] string userNombreUsuario, [FromHeader] string userContraseña)
         {
-            // ... lógica para buscar y eliminar la orden ...
+            var existingUser = _serviceContext.Set<UserItem>()
+       .Where(u => u.NombreUsuario == userNombreUsuario && 
+       u.Contraseña == userContraseña && u.Rol == 1)
+       .FirstOrDefault();
+
+            if (existingUser == null)
+            {
+                return Unauthorized();
+            }
+
+            var existingOrderItem = _serviceContext.Set<OrderItem>()
+                .Where(p => p.IdOrder == id)
+                .FirstOrDefault();
+
+            if (existingOrderItem == null)
+            {
+                return NotFound();
+            }
 
             _orderService.DeleteOrder(id);
             return Ok();
         }
+
+        [HttpGet("{id}", Name = "GetOrderById")]
+        public IActionResult Get(int id)
+        {
+            var orderItem = _orderService.GetOrderById(id); 
+
+            if (orderItem == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(orderItem);
+        }
+
+
     }
+
 }
